+++ src/Application.php
@@ -2,14 +2,16 @@
 
 namespace Rougin\Slytherin;
 
+use Rougin\Slytherin\System\Routing;
+
 /**
  * Application
  *
- * NOTE: To be removed in v1.0.0. Use "System" instead.
+ * NOTE: To be removed in v1.0.0. Use "Routing" instead.
  *
  * @package Slytherin
  * @author  Rougin Gutib <rougingutib@gmail.com>
  */
-class Application extends System
+class Application extends Routing
 {
 }